refactor: Post 목록 분리
