+++ src/features/posts/ui/PostsSearchHeader.tsx
